@@ -8,39 +8,39 @@
   "license": "MIT",
   "private": true,
   "devDependencies": {
-    "@babel/core": "7.14.0",
-    "@babel/plugin-transform-runtime": "7.13.15",
-    "@babel/preset-env": "7.14.1",
-    "@babel/preset-react": "7.13.13",
+    "@babel/core": "7.16.0",
+    "@babel/plugin-transform-runtime": "7.16.4",
+    "@babel/preset-env": "7.16.4",
+    "@babel/preset-react": "7.16.0",
     "@poool/eslint-config": "1.5.0",
     "@poool/eslint-config-react": "1.5.1",
     "@poool/eslint-plugin": "1.5.0",
     "babel-eslint": "10.1.0",
-    "babel-loader": "8.2.2",
-    "eslint": "7.30.0",
+    "babel-loader": "8.2.3",
+    "eslint": "8.3.0",
     "eslint-config-standard": "16.0.3",
     "eslint-plugin-babel": "5.3.1",
-    "eslint-plugin-import": "2.23.4",
+    "eslint-plugin-import": "2.25.3",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "5.1.0",
-    "eslint-plugin-react": "7.24.0",
-    "eslint-plugin-standard": "4.1.0",
+    "eslint-plugin-promise": "5.1.1",
+    "eslint-plugin-react": "7.27.1",
+    "eslint-plugin-standard": "5.0.0",
     "file-loader": "6.2.0",
-    "html-webpack-plugin": "5.3.2",
-    "webpack": "5.36.2",
-    "webpack-cli": "4.7.2",
-    "webpack-dev-server": "3.11.2"
+    "html-webpack-plugin": "5.5.0",
+    "webpack": "5.64.3",
+    "webpack-cli": "4.9.1",
+    "webpack-dev-server": "4.5.0"
   },
   "dependencies": {
-    "@babel/eslint-parser": "7.13.14",
-    "@babel/eslint-plugin": "7.13.16",
-    "@babel/runtime-corejs3": "7.14.0",
+    "@babel/eslint-parser": "7.16.3",
+    "@babel/eslint-plugin": "7.14.5",
+    "@babel/runtime-corejs3": "7.16.3",
     "@poool/react-access": "1.2.0",
-    "core-js": "3.15.2",
-    "history": "5.0.0",
+    "core-js": "3.19.1",
+    "history": "5.1.0",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-router-dom": "5.2.0"
+    "react-router-dom": "6.0.2"
   },
   "scripts": {
     "start": "node_modules/.bin/webpack serve --config ./webpack.config.js",
