@@ -1,8 +1,12 @@
 import React, { useEffect } from 'react';
-import { Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { PaywallContext } from '@poool/react-access';
 
 import { defaultHistory } from '../utils';
-import AppRouter from './AppRouter';
+import Home from './Home';
+import Premium from './Premium';
+import Free from './Free';
+import Subscription from './Subscription';
 
 // Avoid redux
 window.testUser = window.testUser || {
@@ -16,10 +20,18 @@ export default () => {
   }, []);
 
   return (
-    <Router history={defaultHistory}>
-      <Switch>
-        <Route component={AppRouter} />
-      </Switch>
-    </Router>
+    <BrowserRouter history={defaultHistory}>
+      <PaywallContext
+        appId="155PF-L7Q6Q-EB2GG-04TF8"
+        config={{ debug: true, cookies_enabled: true, custom_segment: 'react' }}
+      >
+        <Routes>
+          <Route path="/premium" element={<Premium />} />
+          <Route path="/free" element={<Free />} />
+          <Route path="/subscribe" element={<Subscription />} />
+          <Route index element={<Home />} />
+        </Routes>
+      </PaywallContext>
+    </BrowserRouter>
   );
 };
